websockets: Return from writer thread after close finished
When an interrupt arrived, the writer thread sends a CloseMessage, the reader thread receives confirmation and nicely shuts down, closing the done channel and the websocket. If that happens within one second, the writer thread needs to return rather than continuing in the loop. If that didn't happen within one second, only then it's useful for the writer thread to call c.Close() as well.

The most important fix is changing break to return, to avoid leaking the writer thread and having it fail every second.